I think the GC just passed its first test case!
Basically it allocates (and then throws away) a bunch of really big objects, about twice as many as the space can actually hold. And when it runs out of memory partway through, it does a scavenge and then keeps going.

Now, the scavenge takes about ten minutes (on a 67 MB space), but still...